backend/requirements.txt
django==4.0.3
python-dotenv==0.20.0
djangorestframework==3.13.1
PyJWT==2.3.0
gunicorn==20.1.0
psycopg-binary==3.0.10
djangorestframework-simplejwt==5.1.0
django-colorfield==0.6.3
psycopg2==2.9.3
djoser==2.1.0
webcolors==1.11.1
drf-extra-fields==3.4.0